WW-5453 Rename VelocityManager interface and default implementation
