+++ plugins/sitemesh/src/main/java/org/apache/struts2/sitemesh/OldDecorator2NewStrutsVelocityDecorator.java
@@ -29,7 +29,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.apache.struts2.views.velocity.VelocityManagerInterface;
+import org.apache.struts2.views.velocity.VelocityManager;
 import org.apache.velocity.context.Context;
 
 import java.io.IOException;
@@ -41,10 +41,10 @@ import java.io.PrintWriter;
 public class OldDecorator2NewStrutsVelocityDecorator extends OldDecorator2NewStrutsDecorator {
     private static final Logger LOG = LogManager.getLogger(OldDecorator2NewStrutsFreemarkerDecorator.class);
 
-    private static VelocityManagerInterface velocityManager;
+    private static VelocityManager velocityManager;
 
     @Inject(required = false)
-    public static void setVelocityManager(VelocityManagerInterface mgr) {
+    public static void setVelocityManager(VelocityManager mgr) {
         velocityManager = mgr;
     }
